requirements.txt
streamlit>=1.31.0
duckdb==1.4.4
pandas
plotly
MLB-StatsAPI
python-dotenv
